# openapi.yaml
openapi: 3.0.3
info:
  title: Merch Shop API
  description: E-Sport merch shop REST API
  version: 1.0.0
servers:
  - url: /api
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    CreatePublisherDto:
      type: object
      required: [name, slug]
      properties:
        name:
          type: string
        slug:
          type: string
        logoUrl:
          type: string
          format: uri
    CreateGameDto:
      type: object
      required: [publisherId, name, slug]
      properties:
        publisherId:
          type: string
        name:
          type: string
        slug:
          type: string
        imageUrl:
          type: string
          format: uri
...
    AddToCartDto:
      type: object
      required: [skuId, quantity]
      properties:
        skuId:
          type: string
        quantity:
          type: integer
          minimum: 1
...
    RegisterDto:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 12
    LoginDto:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
    ForgotPasswordDto:
      type: object
      required: [email]
      properties:
        email:
          type: string
          format: email
    ResetPasswordDto:
      type: object
      required: [token, newPassword]
      properties:
        token:
          type: string
        newPassword:
          type: string
          minLength: 12
    VerifyEmailDto:
      type: object
      required: [token]
      properties:
        token:
          type: string
paths:
...
  /auth/register:
    post:
      tags: [auth]
      summary: Register a new buyer account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterDto"
      responses:
        "201":
          description: Created
        "409":
          description: Email already registered
  /auth/login:
    post:
      tags: [auth]
      summary: Log in and receive an access token (refresh token set as an HttpOnly cookie)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginDto"
      responses:
        "200":
          description: OK
        "401":
          description: Invalid credentials
  /auth/refresh:
    post:
      tags: [auth]
      summary: Rotate the refresh-token cookie and issue a new access token
      responses:
        "200":
          description: OK
        "401":
          description: Missing, invalid, or expired refresh token
...
  /auth/forgot-password:
    post:
      tags: [auth]
      summary: Request a password-reset email (always returns a generic response)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ForgotPasswordDto"
      responses:
        "200":
          description: OK
  /auth/reset-password:
    post:
      tags: [auth]
      summary: Reset a password using a reset token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResetPasswordDto"
      responses:
        "200":
          description: OK
        "400":
          description: Invalid or expired token
  /auth/verify-email:
    post:
      tags: [auth]
      summary: Verify an account's email using a verification token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VerifyEmailDto"
      responses:
        "200":
          description: OK
        "400":
          description: Invalid or expired token

  # --- Publishers ---
  /publishers:
    get:
      tags: [publishers]
      summary: List all publishers
      responses:
        "200":
          description: OK
    post:
      tags: [publishers]
      summary: Create publisher (admin)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePublisherDto"
      responses:
        "201":
          description: Created
  /publishers/{slug}:
    get:
      tags: [publishers]
      summary: Get publisher by slug
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
        "404":
          description: Not found
  /publishers/{id}:
    patch:
      tags: [publishers]
      summary: Update publisher (admin)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePublisherDto"
      responses:
        "200":
          description: OK
    delete:
      tags: [publishers]
      summary: Delete publisher (admin)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK

  # --- Games ---
  /games:
    get:
      tags: [games]
      summary: List all games
      parameters:
        - name: publisherId
          in: query
          schema:
            type: string
      responses:
        "200":
          description: OK
    post:
      tags: [games]
      summary: Create game (admin)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateGameDto"
      responses:
        "201":
          description: Created
  /games/{slug}:
    get:
      tags: [games]
      summary: Get game by slug
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
        "404":
          description: Not found
  /games/{id}:
    patch:
      tags: [games]
      summary: Update game (admin)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateGameDto"
      responses:
        "200":
          description: OK
    delete:
      tags: [games]
      summary: Delete game (admin)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK

  # --- Teams ---
  /teams:
    get:
      tags: [teams]
      summary: List all teams
      parameters:
        - name: gameId
          in: query
          schema:
            type: string
      responses:
        "200":
          description: OK
    post:
      tags: [teams]
      summary: Create team (admin)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTeamDto"
      responses:
        "201":
          description: Created
  /teams/{slug}:
    get:
      tags: [teams]
      summary: Get team by slug
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
        "404":
          description: Not found
  /teams/{id}:
    patch:
      tags: [teams]
      summary: Update team (admin)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTeamDto"
      responses:
        "200":
          description: OK
    delete:
      tags: [teams]
      summary: Delete team (admin)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK

  # --- Characters ---
  /characters:
    get:
      tags: [characters]
      summary: List all characters
      parameters:
        - name: gameId
          in: query
          schema:
            type: string
      responses:
        "200":
          description: OK
    post:
      tags: [characters]
      summary: Create character (admin)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCharacterDto"
      responses:
        "201":
          description: Created
  /characters/{slug}:
    get:
      tags: [characters]
      summary: Get character by slug
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
        "404":
          description: Not found
  /characters/{id}:
    patch:
      tags: [characters]
      summary: Update character (admin)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCharacterDto"
      responses:
        "200":
          description: OK
    delete:
      tags: [characters]
      summary: Delete character (admin)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK

  # --- Products ---
  /products:
    get:
      tags: [products]
      summary: List all products
      description: >-
        Public by default. An optional Bearer token is accepted; when it
        belongs to an admin account, `includeUnavailable` takes effect.
      parameters:
        - name: gameId
          in: query
          schema:
            type: string
        - name: teamId
          in: query
          schema:
            type: string
        - name: characterId
          in: query
          schema:
            type: string
        - name: includeUnavailable
          in: query
          description: >-
            Include skus with available: false. Requires an admin Bearer
            token; ignored otherwise.
          schema:
            type: boolean
      responses:
        "200":
          description: OK
    post:
      tags: [products]
      summary: Create product (admin)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProductDto"
      responses:
        "201":
          description: Created
  /products/{id}:
    get:
      tags: [products]
      summary: Get product by id
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
        "404":
          description: Not found
    patch:
      tags: [products]
      summary: Update product (admin)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProductDto"
      responses:
        "200":
          description: OK
    delete:
      tags: [products]
      summary: Delete product (admin)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK

  # --- SKUs ---
  /skus:
    get:
      tags: [skus]
      summary: List SKUs by product
      parameters:
        - name: productId
          in: query
          schema:
            type: string
      responses:
        "200":
          description: OK
    post:
      tags: [skus]
      summary: Create SKU (admin)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSkuDto"
      responses:
        "201":
          description: Created
  /skus/{id}/availability:
    patch:
      tags: [skus]
      summary: Set SKU availability (admin)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [available]
              properties:
                available:
                  type: boolean
      responses:
        "200":
          description: OK
  /skus/availability/bulk:
    patch:
      tags: [skus]
      summary: Set availability for all SKUs under a facet (admin)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BulkAvailabilityDto"
      responses:
        "200":
          description: OK
  /skus/{id}:
    delete:
      tags: [skus]
      summary: Delete SKU (admin)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK

  # --- Cart ---
...
  /cart/items:
    post:
      tags: [cart]
      summary: Add item to cart (cookie or token)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddToCartDto"
      responses:
        "201":
          description: Created
  /cart/items/{skuId}:
    delete:
      tags: [cart]
      summary: Remove item from cart (cookie or token)
      parameters:
        - name: skuId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
  /cart/sync:
    post:
      tags: [cart]
      summary: Replace authenticated user's cart contents
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SyncCartDto"
      responses:
        "201":
          description: Synced
  /cart/merge:
    post:
      tags: [cart]
      summary: Merge guest cart into authenticated user cart
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [sessionId]
              properties:
                sessionId:
                  type: string
      responses:
        "201":
          description: Merged

  # --- Orders ---
  /orders:
    get:
      tags: [orders]
      summary: List all orders (admin)
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: status
          in: query
          schema:
            type: string
            enum: [PENDING, CONFIRMED, FORWARDED, CANCELLED]
      responses:
        "200":
          description: OK
  /orders/mine:
    get:
      tags: [orders]
      summary: List authenticated user's orders
      security:
        - bearerAuth: []
      responses:
        "200":
          description: OK
  /orders/{id}:
    get:
      tags: [orders]
      summary: Get order by id
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
        "404":
          description: Not found
  /orders/by-payment-intent/{intentId}:
    get:
      tags: [orders]
      summary: Get authenticated user's order by Stripe payment intent id
      security:
        - bearerAuth: []
      parameters:
        - name: intentId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
        "404":
          description: Not found
  /orders/{id}/retry-fulfillment:
    post:
      tags: [orders]
      summary: Retry fulfillment forwarding for an order (admin)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK

  # --- Account ---
  /account/me:
    get:
      tags: [account]
      summary: Get or create account for authenticated user
      security:
        - bearerAuth: []
      responses:
        "200":
          description: OK
  /account/{id}:
    delete:
      tags: [account]
      summary: Delete account (admin)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK

  # --- Payments ---
  /payments/payment-intent:
    post:
      tags: [payments]
      summary: Create Stripe PaymentIntent
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [cartId]
              properties:
                cartId:
                  type: string
      responses:
        "201":
          description: PaymentIntent created
...